refactor: roomService 로 room 관련된 비지니스 로직 분리
